requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
greenlet==2.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pendulum==2.1.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytzdata==2020.1
six==1.16.0
SQLAlchemy==1.4.44
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
